[MOVE/#14] bottom sheet 위치 수정
