+++ core/src/main/res/drawable/ic_character1.xml
