+++ core/src/main/res/drawable/ic_character2.xml
