+++ core/src/main/res/drawable/ic_character3.xml
