+++ core/src/main/res/drawable/ic_character4.xml
