+++ core/src/main/res/drawable/ic_character5.xml
