+++ core/src/main/res/drawable/ic_character6.xml
