+++ core/src/main/res/drawable/ic_selected_character.xml
