+++ feature/src/main/java/com/terning/feature/main/MainNavigator.kt
@@ -13,6 +13,7 @@ import com.terning.feature.calendar.navigation.navigateCalendar
 import com.terning.feature.home.navigation.navigateHome
 import com.terning.feature.mypage.navigation.navigateMyPage
 import com.terning.feature.onboarding.signin.navigation.SignIn
+import com.terning.feature.onboarding.signup.navigation.SignUp
 import com.terning.feature.search.search.navigation.navigateSearch
 
 class MainNavigator(
@@ -22,7 +23,7 @@ class MainNavigator(
         @Composable get() = navController
             .currentBackStackEntryAsState().value?.destination
 
-    val startDestination = SignIn
+    val startDestination = SignUp
 
     val currentTab: MainTab?
         @Composable get() = MainTab.find { tab ->
